sik-1406
change query date range for track points so we dont always publish all of them.
move condition for checking if we skip processing to bottom after we determine if we have found new track points. Previously we would skip patrol processing before determining if there were new track points as those are not included in the updates array