Dropdown fixes
Improvements to the add device form and edit device group form. By changing the str methods for OutboundIntegrationConfiguration and InboundIntegrationConfiguration, the dropdowns will now be sorted by the Owners of the integrations. This makes it easier for the user to comprehend their dropdown choices. Also the help_text for the destination field on edit device group has been expanded